refactor/#7: BootstrapService를 추가하여 Swagger 설정 로직을 분리
