+++ src/main.ts
@@ -1,43 +1,18 @@
 import { Logger } from "@nestjs/common";
 import { NestFactory } from "@nestjs/core";
-import { DocumentBuilder, SwaggerModule } from "@nestjs/swagger";
 import { AppModule } from "@src/app.module";
+import { BootstrapService } from "@src/bootstrap.service";
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule, {
     bufferLogs: true,
   });
+  const bootstrapService = app.get<BootstrapService>(BootstrapService);
+
+  bootstrapService.setSwagger(app);
 
   app.useLogger(app.get(Logger));
 
-  // Swagger 설정
-  const config = new DocumentBuilder()
-    .setTitle("Sift API")
-    .setDescription("Sift 백엔드 API 문서입니다.")
-    .setVersion("1.0")
-    .addBearerAuth()
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup("api-docs", app, document, {
-    swaggerOptions: {
-      persistAuthorization: true,
-      targetSorter: "alpha",
-      operationsSorter: (a: Map<any, any>, b: Map<any, any>) => {
-        const methodsOrder = {
-          post: "0",
-          get: "1",
-          put: "2",
-          patch: "3",
-          delete: "4",
-        };
-
-        return methodsOrder[a.get("method")].localeCompare(
-          methodsOrder[b.get("method")],
-        );
-      },
-    },
-  });
-
-  await app.listen(process.env.PORT || 3000);
+  await bootstrapService.startingServer(app);
 }
 bootstrap();
